Move `_module_extra_repr` to `torchao/quantization/utils.py`
Summary:
Moving to utils.py so we can use this both in non-prototype and prototype quant_api.py

Test Plan:
CI

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
